refactor(driver): database connection code refactored
